+++ docs/changelog.md
@@ -2,6 +2,8 @@
 
 # Version 3.3.3
 
+- *Crosshairs* - Added `Sequencer.Crosshair.CALLBACKS.STOP_COLLIDING` callback
+- *Crosshairs* - Added `updateCrosshair` method to crosshairs
 - *Crosshairs* - Fixed invalid placement positions being cached and then used even if the crosshair was placed somewhere else
 - *Effects* - Fixed `.attachTo()`'s `align` not playing well with `.stretchTo()`
 - *Effects* - Fixed `.persist(true, { persistTokenPrototype: true })` not storing more than one persistent effect on the token